apartment_agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.28.1
apscheduler>=3.10.4
pydantic>=2.9.2
crewai[litellm]>=0.175.0
